% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EstimatesTable.R
\name{EstimatesTable}
\alias{EstimatesTable}
\title{Create Mplus Estimates Table}
\usage{
EstimatesTable(
  ModelsIn,
  std = "unstd",
  ci = F,
  ci.limit = "",
  one.path.col = T,
  show.p = F,
  show.sign = T,
  alpha = c(0.001, 0.01, 0.05),
  decimals = 2,
  p.decimals = 3
)
}
\arguments{
\item{ModelsIn}{List of one or more model objects resulting from MplusAutomation::readModels}

\item{std}{Vector of one to four standardization methods to be included ("unstd", "stdyx", "stdy" or "std").}

\item{ci}{Display confidence intervals? Default is False, meaning se and p values will be displayed.}

\item{ci.limit}{Which confidence interval limits to display ("90", "95" or "99").}

\item{one.path.col}{If true, path is displayed as one column. If false, three path columns are created (X, Direction and Y)}

\item{show.p}{Display p-values? If TRUE, p-values will be presented in a separate column}

\item{show.sign}{Display significance? If TRUE, significance asterisks will be displayed with the estimates.}

\item{alpha}{A vector of one to three numbers in descending order to assess significance levels.}

\item{decimals}{A single number indicating to how many decimals should be rounded ( applies to all numbers except p-values)}

\item{p.decimals}{A single number indicating to how many decimals p values should be rounded.}
}
\description{
This function takes a list of models as input (in which each list item
is the result of an mplus model), and creates a table for all models'
estimates. (Each model is below the previous.)
}
\details{
The standardization methods (std) are based on those exported by Mplus. The options are:
(1) "unstd": provides unstandardized estimates;
(2) "stdyx": provides stdyx-standardized estimates;
(3) "stdy": provides stdy-standardized estimates;
(4)"std": provides std-standardized estimates.
By default, Mplus output only shows unstandardized estimates. Use the command "Output: standardized;"
in Mplus to request all three standardization methods. Note that if you enter an invalid standardization method,
or request standardized estimates where there were none in the original output file, you will receive an error.

The options for confidence intervals are based on those exported by Mplus. The options are:
90\%, 95\% or 99\% confidence intervals. By default, Mplus output does not show confidence intervals.
Use the command "Output: cinterval;" in Mplus to request confidence intervals. Note that if you
enter invalid limits, or request confidence intervals where there were none in the original output file,
you will receive an error.
}
\examples{
M1output <- readModels("M1.out") # Uses the MplusAutomation package
M1estimates <- EstimatesTable(M1output)

# Create a table with stdyx-standardized estimates and p values in a separate column
M1estimates <- EstimatesTable(M1output, std = "stdyx", show.p = T, show.sign = F)

# Create a table with unstandardized and standardized estimates
M1estimates <- EstimatesTable(M1output, std = c("unstd","stdyx"))

# Create a table with 95\% confidence intervals
M1estimates <- estimatesTable(M1output, std="unstd", ci=T, ci.limit="95")

}
